Optimized get all reference rate functionalities

+++ code/src/main/java/com/example/formedix/controllers/MainController.java
@@ -26,7 +26,7 @@ public class MainController {
     public ResponseEntity<Map<String, Object>> getReferenceRate(@RequestParam(value = "date", required = true) String date){
         response.clear();
         try {
-            Map<String, Float> result = mainService.getRates(date);
+            Map<String, Double> result = mainService.getRates(date);
             response.put("status","success");
             response.put("response", result);
 
@@ -97,5 +97,5 @@ public class MainController {
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
-    
+
 }
