+++ code/src/main/java/com/example/formedix/services/MainService.java
@@ -4,7 +4,7 @@ import com.example.formedix.exceptions.CustomException;
 import java.util.Map;
 
 public interface MainService {
-    Map<String, Float> getRates(String date) throws CustomException;
+    Map<String, Double> getRates(String date) throws CustomException;
 
     double convertCurrency(String dates, String source, String target, String amount) throws CustomException;
